Move SearchControl to controls folder
